+++ agrest-engine/src/main/java/io/agrest/DeleteBuilder.java
@@ -2,8 +2,12 @@ package io.agrest;
 
 import io.agrest.access.DeleteAuthorizer;
 import io.agrest.meta.AgEntityOverlay;
+import io.agrest.processor.Processor;
+import io.agrest.processor.ProcessorOutcome;
+import io.agrest.runtime.processor.delete.DeleteContext;
 
 import java.util.Map;
+import java.util.function.Consumer;
 
 /**
  * @since 1.4
@@ -45,6 +49,58 @@ public interface DeleteBuilder<T> {
      * @since 4.8
      */
     DeleteBuilder<T> authorizer(DeleteAuthorizer<T> authorizer);
+
+    /**
+     * Registers a consumer to be executed after a specified standard execution stage. The consumer can inspect and
+     * modify provided {@link DeleteContext}.
+     * <p>This operation is composable. For each stage all custom processors will be invoked in the order they were
+     * registered.</p>
+     *
+     * @param afterStage  A name of the standard stage after which the inserted stage needs to be run.
+     * @param customStage a callback to invoke at a specific point during  the update execution.
+     * @return this builder instance.
+     * @since 4.8
+     */
+    default <U> DeleteBuilder<T> stage(DeleteStage afterStage, Consumer<DeleteContext<U>> customStage) {
+        return routingStage(afterStage, (DeleteContext<U> c) -> {
+            customStage.accept(c);
+            return ProcessorOutcome.CONTINUE;
+        });
+    }
+
+    /**
+     * Registers a consumer to be executed after the specified standard execution stage. The rest of the standard pipeline
+     * following the named stage will be skipped. The consumer can inspect and modify provided {@link DeleteContext}.
+     * <p>This operation is composable. For each stage all custom processors will be invoked in the order they were
+     * registered.</p>
+     *
+     * @param afterStage          A name of the standard stage after which the inserted stage needs to be run.
+     * @param customTerminalStage a consumer that will be invoked after 'afterStage', and will be the last piece of
+     *                            code executed in the update pipeline.
+     * @return this builder instance.
+     * @since 4.8
+     */
+    default <U> DeleteBuilder<T> terminalStage(DeleteStage afterStage, Consumer<DeleteContext<U>> customTerminalStage) {
+        return routingStage(afterStage, (DeleteContext<U> c) -> {
+            customTerminalStage.accept(c);
+            return ProcessorOutcome.STOP;
+        });
+    }
+
+    /**
+     * Registers a processor to be executed after the specified standard execution stage. The processor can inspect and
+     * modify provided {@link DeleteContext}. When finished, processor can either pass control to the next stage by returning
+     * {@link ProcessorOutcome#CONTINUE}, or terminate the pipeline by returning
+     * {@link ProcessorOutcome#STOP}.
+     * <p>This operation is composable. For each stage all custom processors will be invoked in the order they were
+     * registered.</p>
+     *
+     * @param afterStage  A name of the standard stage after which the inserted stage needs to be run.
+     * @param customStage a processor to invoke at a specific point during the update execution.
+     * @return this builder instance.
+     * @since 4.8
+     */
+    <U> DeleteBuilder<T> routingStage(DeleteStage afterStage, Processor<DeleteContext<U>> customStage);
 
     SimpleResponse delete();
 }