+++ agrest-engine/src/main/java/io/agrest/runtime/DefaultDeleteBuilder.java
@@ -3,14 +3,17 @@ package io.agrest.runtime;
 import io.agrest.AgException;
 import io.agrest.AgObjectId;
 import io.agrest.DeleteBuilder;
+import io.agrest.DeleteStage;
 import io.agrest.EntityParent;
 import io.agrest.SimpleResponse;
 import io.agrest.access.DeleteAuthorizer;
 import io.agrest.meta.AgEntity;
 import io.agrest.meta.AgEntityOverlay;
+import io.agrest.processor.Processor;
 import io.agrest.runtime.processor.delete.DeleteContext;
 import io.agrest.runtime.processor.delete.DeleteProcessorFactory;
 
+import java.util.EnumMap;
 import java.util.Map;
 
 /**
@@ -18,12 +21,14 @@ import java.util.Map;
  */
 public class DefaultDeleteBuilder<T> implements DeleteBuilder<T> {
 
-    protected DeleteContext<T> context;
-    protected DeleteProcessorFactory processorFactory;
+    protected final DeleteContext<T> context;
+    protected final DeleteProcessorFactory processorFactory;
+    protected final EnumMap<DeleteStage, Processor<DeleteContext<?>>> processors;
 
     public DefaultDeleteBuilder(DeleteContext<T> context, DeleteProcessorFactory processorFactory) {
         this.context = context;
         this.processorFactory = processorFactory;
+        this.processors = new EnumMap<>(DeleteStage.class);
     }
 
     @Override
@@ -75,8 +80,18 @@ public class DefaultDeleteBuilder<T> implements DeleteBuilder<T> {
     }
 
     @Override
+    public <U> DeleteBuilder<T> routingStage(DeleteStage afterStage, Processor<DeleteContext<U>> customStage) {
+        return routingStage_NoGenerics(afterStage, customStage);
+    }
+
+    private DeleteBuilder<T> routingStage_NoGenerics(DeleteStage afterStage, Processor customStage) {
+        processors.compute(afterStage, (s, existing) -> existing != null ? existing.andThen(customStage) : customStage);
+        return this;
+    }
+
+    @Override
     public SimpleResponse delete() {
-        processorFactory.createProcessor().execute(context);
+        processorFactory.createProcessor(processors).execute(context);
         return context.createSimpleResponse();
     }
 }